backend/requirements.txt
# FastAPI & Web
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.5.3
pydantic-settings==2.1.0

# OpenAI & LangChain
openai==1.12.0
langchain==0.1.6
langchain-openai==0.0.5
langgraph==0.0.26

# Vector Database
qdrant-client==1.7.0

# Document Parsing
PyMuPDF==1.23.8
python-docx==1.1.0

# Observability
langfuse==2.9.0

# Database
sqlalchemy==2.0.25
aiosqlite==0.19.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0

# Utilities
python-dotenv==1.0.0
httpx==0.26.0
aiofiles==23.2.1
